Split release modules defs
The releases option set by various CUDA libraries (cudnn, tensorrt,
etc.) share a common base but each adds a few different options. This
commit factors the commonalities in a separate submodule, that is then
used (in conjunction with `submoduleWith`) to combine an extension and
generate the corresponding config options without code duplication.