Enable cuda_compat by default on supported devices
Some nvidia devices, such as the Jetson family, support a package called
nvidia compatibility package (nvidia_compat) which allows to run
executables built against a higher CUDA major version on a system with
an older CUDA driver. On the platform, the consensus among CUDA
maintainers is that there is no downside in always enabling it by default.

This commits link to the relevant cuda_compat shared libraries by
patching the executables with the relevant runpaths when cuda_compat is
available, in the same way as we do for OpenGL drivers currently.